complete tests for asgi.py
